Extracted find machine files funcs into new script
